+++ CHANGES.md
@@ -14,11 +14,13 @@ Core Changes:
 
 Language Improvements:
 
-- none.
+- enh(plaintext) added `text` and `txt` as alias (#2360) [Taufik Nurrohman][]
 
 Developer Tools:
 
 - none.
+
+[Taufik Nurrohman]: https://github.com/taufik-nurrohman
 
 
 ## Version 9.18.0